Fix __exit__() test
This patch fixes tests for proxying the __exit__ method in case an
exception is actually raised. The original test didn't test the
traceback and actually didn't test anything at all, due to the missing
call to the assertion method. Now the original exception traceback is
collected and compared (and it all works)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>