Rename unproxied to direct, proxiee to original and move them to proxy cls
This patch, which could have been argualby broken up but I didn't want
to spend more time on this than strictly required does the following
operations:

- renames padme.unproxied to padme.proxy.direct
- renames padme.proxiee to padme.proxy.original
- padme.__all__ no longer lists unproxied or proxiee
- various references are updated to use the new names
- a small new documentation section mentions the deprecated APIs

This is done so that the names are a little bit easier to understand
and so that applications need to import just one thing from padme (the
proxy class itself). Both functions are just @staticmethods defined on
the class for pure convenience.

The old functions are replaced with simple aliases for full backward
compatibility.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>