Use _get_unproxied to make proxy_base agnostic of how proxy itself works
I want to switch to a totally different implementation of base proxy
concept and to do that I want to share the proxy_base class by making
it unaware of how to access the set of direct-access attributes.
The new temporary function _get_unproxied() makes that easier to change
later.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>