Split statful_proxy_meta off proxy_meta
This patch adds a new subclass of proxy_meta, the stateful_proxy_meta
class, which carries the __proxy_state__ attribute that used to exist
on the base proxy class. The goal is to remove any instance-specific
data from the base classes, in anticipation of the big type split later.

All uses of proxy_meta were converted to use stateful_proxy_meta instead

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>