Add make_typed_proxy_meta()
This patch adds a new support function, mimicking make_boundproxy_meta(),
which creates a new subclass of proxy_meta() that is not related to the
proxiee in any way. Instead it is related to the type of proxiee. This
allows all such metaclasses (and later on actual proxy classes) shared
per type so that ``type(proxy(42) is type(proxy(24))`` will finally
evaulate to True.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>