Switch to instance-decoupled metaclasses
This large patch makes proxies more light-weight than before by decoupling
the both the class and the metaclass of each proxy decoupled from the
proxieed object.

Before, each time a proxy class was used to create a proxy object it would
transparently create a new subclass of itself. The new subclass would be
defined based on a new meta-class that would wrap the proxieed object
itself.

In graphical / ascii-art terms it would look something like this:
In the diagram, arrows pointing up are represent inheritance while
arrows pointing right represent meta-classes.
 +------------+
 | proxy_base |
 +------------+
        ^
        |
 +------------+       +-------------+
 |   proxy    |  ---> | proxy_meta  |
 +------------+       +-------------+
       ^                     ^
       |                     |
 +------------+       +-----------------+
 | proxy      |       | boundproxy_meta |
 | unique     |  ---> | unique to       |
 | to proxiee |       | to proxiee      |
 +------------+       +-----------------+

This scheme was needlessly costly. Meta-classes are just used so that
__subclasscheck__ can correctly "lie" about the type of the proxy.

In is obvious, that for any instance of a proxy subclass (including the
proxy class itself), wrapping the original object of type T, only one
metaclass, ;et's call it proxy_meta[T], and one concrete proxy class,
let's call it proxy[T], is needed.

The type-specialized metaclass is always shared among proxy subclasses.
The type-specialized class is shared within that class as subclasses of
proxy may define a different set of __unproxied__ methods. In any case
the concrete object returned by proxy_cls(original) must lookup or create
a proxy class that uses proxy_meta[T] as its meta class and proxy_cls[T]
as its base class.

The registry of proxy_meta[T] classes is stored in proxy._registry_m.
The registry of proxy[T] classes is stored in proxy._registry_c and is
also reset to {} for each subclass using proxy_meta as its metaclass
(again, so that __unproxied__ can be computed correctly).

This sharing also means that we cannot use type(proxy_obj).__proxiee__
as a way to store the reference to the proxied original object. Instead,
the state of each proxy is used to store the original object as _original.

In plain terms, proxy.state(proxy_obj)._original is always the original
object.

I haven't done any benchmarks on this but the new scheme is more natural
and definitely has less references to unique types than the old one.
For completeness, this is the diagram of inheritance for the new scheme:

 +------------+
 | proxy_base |
 +------------+
     ^
     |
 +-------+       +------------+
 | proxy |  ---> | proxy_meta |
 +-------+       +------------+
     ^                  ^
     |                  |
 +----------+     +---------------+
 | proxy[T] | --> | proxy_meta[T] |
 +----------+     +---------------+

Here, the obvious big difference is, again, total de-coupling of the type
system from the *object* being proxied. Only the type of the object being
proxied is important.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>