Redirect all __iFUNC__ methods to the proxy if they exist
This change is a bit interesting. The way augmented assignment works
we cannot just let calls to __iadd__ and friends fall through to the
proxiee. This would cause a simple snippet like this to fail:

a = proxy(2)
a += 2
assert issubclass(type(a), proxy)

This is because python internally replaces ``a`` there with the result
of the addition. Here int is also extra interesting as it doesn't have
the __iadd__ method so ``a += 2`` becomes ``a = a + 2``. Since we want
to keep being the proxy object, all such methods are redirected to the
proxy and handled there.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>