Fix augmented assignment methods on immutable objects
This patch fixes an issue with all augmented assignment methods that was
discovered when doing experiments. A proxy of an immutable type would
behave differently from the immutable value when used together with
augmented assignment methods as the old code (which is fixed by this
patch) would re-write the original object instead of returning a new
value.

In practical terms this test would fail:

 a = b = proxy(1)
 a += 2
 assert a == 3  # this would work
 assert b == 1  # this would fail without this patch

A number of tests like this were added to ensure the new implementation
behaves correctly for commonly used types.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>